+++ CHANGELOG.md
@@ -19,6 +19,9 @@ instance of `GuzzleHttp\Exception\RequestException`. This extends
 without any necessary changes. Errors returned by Maxemail as a response to an
 API call will now throw `Mxm\Api\Exception\ClientException` which extends the
 same `\RuntimeException` used previously.
+- `Helper::setLogLevel()` can be used to increase the log level from the default
+of *debug*. This might be useful if you want to see the file transfer logs at
+*info* and filter out the raw HTTP client logs at *debug*.
 ### Removed
 - Removed support for PHP 5.x as it is no longer
 [actively supported](https://php.net/supported-versions.php) by the PHP project